backend/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0

fastapi==0.128.0
uvicorn[standard]>=0.34.0
python-multipart==0.0.21
python-dotenv==1.0.1
chatterbox-tts>=0.1.6

prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.0.0
